Add JsonRpcHandler to allow more code to be shared between opflex and OVSDB
