+++ libopflex/Makefile.am
@@ -88,6 +88,7 @@ c_include_HEADERS = \
 rpc_includedir = $(includedir)/opflex/rpc
 rpc_include_HEADERS = \
 	include/opflex/rpc/JsonRpcConnection.h \
+	include/opflex/rpc/JsonRpcHandler.h \
 	include/opflex/rpc/JsonRpcMessage.h
 gbp_includedir = $(includedir)/opflex/gbp
 gbp_include_HEADERS = \
